Validate contracts on push instead of add (#224)
Moves all validations to lib package, and centralizes them in a
single validate function. This includes storage layout, constructors,
initializers, delegatecalls, and selfdestructs.

All logging is now centralized in a ValidationLogger class in the CLI,
tested separately from the commands and the validations themselves.

From a user perspective, push will now report all validation errors
and warns, and abort execution if any is found, prompting the user
to run the command with --force to actually proceed. All existing
validation errors are stored in the zos network json file, and only
new validation errors are shown to the user on the next push. These
diffs are not granular enough though: for instance, a contract with
a selfdestruct call will not prompt a new warn if a new selfdestruct
call is added. I'll be adding a new issue for this.

Ideally, these validations should be prompted when the user runs
zos add *if* we used a git-like stage that requires add for staging
changes, but for now, push is the closest thing we have. This has the
disadvantage of tracking errors per-network, where they are actually
dependent on the contract versions.

Finally, these changes make it *really* easy to add a new zos check
command. Let me know if I should work on it.